// apps/web/src/generated/api/masters/masters.ts
export type getMastersRelationshipsResponse200 = {
  data: RelationshipMastersSuccessResponse
  status: 200
}

export type getMastersRelationshipsResponse500 = {
  data: ErrorResponse
  status: 500
}

export type getMastersRelationshipsResponse503 = {
  data: ErrorResponse
  status: 503
}

export type getMastersRelationshipsResponse504 = {
  data: ErrorResponse
  status: 504
}

export type getMastersRelationshipsResponseSuccess = (getMastersRelationshipsResponse200) & {
  headers: Headers;
};
export type getMastersRelationshipsResponseError = (getMastersRelationshipsResponse500 | getMastersRelationshipsResponse503 | getMastersRelationshipsResponse504) & {
  headers: Headers;
};

export type getMastersRelationshipsResponse = (getMastersRelationshipsResponseSuccess | getMastersRelationshipsResponseError)
// ...
export type getMastersOccasionsResponse200 = {
  data: OccasionMastersSuccessResponse
  status: 200
}

export type getMastersOccasionsResponse500 = {
  data: ErrorResponse
  status: 500
}

export type getMastersOccasionsResponse503 = {
  data: ErrorResponse
  status: 503
}

export type getMastersOccasionsResponse504 = {
  data: ErrorResponse
  status: 504
}

export type getMastersOccasionsResponseSuccess = (getMastersOccasionsResponse200) & {
  headers: Headers;
};
export type getMastersOccasionsResponseError = (getMastersOccasionsResponse500 | getMastersOccasionsResponse503 | getMastersOccasionsResponse504) & {
  headers: Headers;
};

export type getMastersOccasionsResponse = (getMastersOccasionsResponseSuccess | getMastersOccasionsResponseError)
// ...
export type getSemanticConfigMastersResponse200 = {
  data: SemanticConfigMastersSuccessResponse
  status: 200
}

export type getSemanticConfigMastersResponse400 = {
  data: ErrorResponse
  status: 400
}

export type getSemanticConfigMastersResponse500 = {
  data: ErrorResponse
  status: 500
}

export type getSemanticConfigMastersResponse503 = {
  data: ErrorResponse
  status: 503
}

export type getSemanticConfigMastersResponseSuccess = (getSemanticConfigMastersResponse200) & {
  headers: Headers;
};
export type getSemanticConfigMastersResponseError = (getSemanticConfigMastersResponse400 | getSemanticConfigMastersResponse500 | getSemanticConfigMastersResponse503) & {
  headers: Headers;
};

export type getSemanticConfigMastersResponse = (getSemanticConfigMastersResponseSuccess | getSemanticConfigMastersResponseError)
// ...
export type getFeatureRuleMastersResponse200 = {
  data: FeatureRuleMastersSuccessResponse
  status: 200
}

export type getFeatureRuleMastersResponse400 = {
  data: ErrorResponse
  status: 400
}

export type getFeatureRuleMastersResponse500 = {
  data: ErrorResponse
  status: 500
}

export type getFeatureRuleMastersResponse503 = {
  data: ErrorResponse
  status: 503
}

export type getFeatureRuleMastersResponseSuccess = (getFeatureRuleMastersResponse200) & {
  headers: Headers;
};
export type getFeatureRuleMastersResponseError = (getFeatureRuleMastersResponse400 | getFeatureRuleMastersResponse500 | getFeatureRuleMastersResponse503) & {
  headers: Headers;
};

export type getFeatureRuleMastersResponse = (getFeatureRuleMastersResponseSuccess | getFeatureRuleMastersResponseError)